Add confirmation dialog and 404 error handling; update styles and routing

+++ TP3/src/app/voyage/components/voyage-list-item/voyage-list-item.component.ts
@@ -3,13 +3,17 @@ import {Voyage} from "../../models/voyage";
 import {CurrencyPipe, SlicePipe} from "@angular/common";
 import {Router} from "@angular/router";
 import {VoyageService} from "../../services/voyage.service";
+import {ConfirmDialogComponent} from "../confirm-dialog/confirm-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
+import {MatButton} from "@angular/material/button";
 
 @Component({
   selector: 'app-voyage-list-item',
   standalone: true,
   imports: [
     SlicePipe,
-    CurrencyPipe
+    CurrencyPipe,
+    MatButton
   ],
   templateUrl: './voyage-list-item.component.html',
   styleUrl: './voyage-list-item.component.css'
@@ -19,15 +23,24 @@ export class VoyageListItemComponent {
   @Output() deleteVoyage = new EventEmitter<number>();
 
 
-  constructor(private router: Router, private voyageService: VoyageService) {
+  constructor(private dialog: MatDialog, private router: Router, private voyageService: VoyageService) {
   }
 
   onSelect() {
     this.router.navigateByUrl('/singleVoyage/' + this.voyage.id);
   }
 
-  onDelete() {
-    this.voyageService.deleteVoyage(this.voyage.id);
-    this.voyageService.getAllVoyages();
+  onDeleteVoyage() {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: {message: 'Are you sure you want to delete this voyage?'}
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+        if (result) {
+            this.voyageService.deleteVoyage(this.voyage.id);
+            this.deleteVoyage.emit(this.voyage.id);
+        }
+    })
   }
 }
